+++ cli/commands/root.go
@@ -31,6 +31,11 @@ const (
 	wunderctlBinaryPathEnvKey = "WUNDERCTL_BINARY_PATH"
 
 	defaultNodeGracefulTimeoutSeconds = 10
+
+	// Telemetry requires the configuration to be generated, so
+	// we need  need to wait a bit for it
+	telemetryReloadWait     = 100 * time.Millisecond
+	maxTelemetryLoadRetries = 20
 )
 
 var (
@@ -55,6 +60,80 @@ var (
 	cyan   = color.New(color.FgHiCyan)
 	white  = color.New(color.FgHiWhite)
 )
+
+func sendTelemetry(cmd *cobra.Command) error {
+	// Check if we want to track telemetry for this command
+	if !rootFlags.Telemetry || !telemetry.HasAnnotations(telemetry.AnnotationCommand, cmd.Annotations) {
+		return nil
+	}
+	cmdMetricName := telemetry.CobraFullCommandPathMetricName(cmd)
+
+	metricDurationName := telemetry.DurationMetricSuffix(cmdMetricName)
+	cmdDurationMetric = telemetry.NewDurationMetric(metricDurationName)
+
+	metricUsageName := telemetry.UsageMetricSuffix(cmdMetricName)
+	cmdUsageMetric := telemetry.NewUsageMetric(metricUsageName)
+
+	metrics := []*telemetry.Metric{cmdUsageMetric}
+
+	wunderGraphDir, err := files.FindWunderGraphDir(_wunderGraphDirConfig)
+	if err != nil {
+		return err
+	}
+
+	var clientInfo *telemetry.MetricClientInfo
+	retries := 0
+	for {
+		clientInfo, err = telemetry.NewClientInfo(BuildInfo.Version, viper.GetString("anonymousid"), wunderGraphDir)
+		if err != nil {
+			if os.IsNotExist(err) {
+				// Configuration is not yet available, wait a bit to try to report all telemetry
+				if retries < maxTelemetryLoadRetries {
+					retries++
+					time.Sleep(telemetryReloadWait)
+					continue
+				}
+			}
+			return err
+		}
+		break
+	}
+
+	// Check if this command should also other telemetry data
+	if telemetry.HasAnnotations(telemetry.AnnotationDataSources, cmd.Annotations) || telemetry.HasAnnotations(telemetry.AnnotationFeatures, cmd.Annotations) {
+		if telemetry.HasAnnotations(telemetry.AnnotationDataSources, cmd.Annotations) {
+			dataSourcesMetrics, err := telemetry.DataSourceMetrics(wunderGraphDir)
+			if err != nil {
+				if rootFlags.TelemetryDebugMode {
+					log.Error("could not generate data sources telemetry data", zap.Error(err))
+
+				}
+			}
+			metrics = append(metrics, dataSourcesMetrics...)
+		}
+
+		if telemetry.HasAnnotations(telemetry.AnnotationFeatures, cmd.Annotations) {
+			featureMetrics, err := telemetry.FeatureMetrics(wunderGraphDir)
+			if err != nil {
+				if rootFlags.TelemetryDebugMode {
+					log.Error("could not generate features telemetry data", zap.Error(err))
+
+				}
+			}
+			metrics = append(metrics, featureMetrics...)
+		}
+	}
+
+	TelemetryClient = telemetry.NewClient(
+		viper.GetString("API_URL"), clientInfo,
+		telemetry.WithTimeout(3*time.Second),
+		telemetry.WithLogger(log),
+		telemetry.WithDebug(rootFlags.TelemetryDebugMode),
+		telemetry.WithAuthToken(os.Getenv("WG_TELEMETRY_AUTH_TOKEN")),
+	)
+
+	return TelemetryClient.Send(metrics)
+}
 
 // rootCmd represents the base command when called without any subcommands
 var rootCmd = &cobra.Command{
@@ -106,62 +185,19 @@ var rootCmd = &cobra.Command{
 			)
 		}
 
-		// Check if we want to track telemetry for this command
-		if rootFlags.Telemetry && telemetry.HasAnnotations(telemetry.AnnotationCommand, cmd.Annotations) {
-			cmdMetricName := telemetry.CobraFullCommandPathMetricName(cmd)
+		// Send telemetry in a goroutine to not block the command
+		go func() {
+			err := sendTelemetry(cmd)
 
-			metricDurationName := telemetry.DurationMetricSuffix(cmdMetricName)
-			cmdDurationMetric = telemetry.NewDurationMetric(metricDurationName)
-
-			metricUsageName := telemetry.UsageMetricSuffix(cmdMetricName)
-			cmdUsageMetric := telemetry.NewUsageMetric(metricUsageName)
-
-			metrics := []*telemetry.Metric{cmdUsageMetric}
-
-			clientInfo := &telemetry.MetricClientInfo{
-				WunderctlVersion: BuildInfo.Version,
-				IsCI:             os.Getenv("CI") != "" || os.Getenv("ci") != "",
-				AnonymousID:      viper.GetString("anonymousid"),
-			}
-
-			// Check if this command should also send data source related telemetry
-			if telemetry.HasAnnotations(telemetry.AnnotationDataSources, cmd.Annotations) {
-				wunderGraphDir, err := files.FindWunderGraphDir(_wunderGraphDirConfig)
+			// AddMetric the usage of the command immediately
+			if rootFlags.TelemetryDebugMode {
 				if err != nil {
-					return err
+					log.Error("Could not send telemetry data", zap.Error(err))
+				} else {
+					log.Info("Telemetry data sent")
 				}
-				dataSourcesMetrics, err := telemetry.DataSourceMetrics(wunderGraphDir)
-				if err != nil {
-					if rootFlags.TelemetryDebugMode {
-						log.Error("could not generate data sources telemetry data", zap.Error(err))
-
-					}
-				}
-				metrics = append(metrics, dataSourcesMetrics...)
 			}
-
-			TelemetryClient = telemetry.NewClient(
-				viper.GetString("API_URL"), clientInfo,
-				telemetry.WithTimeout(3*time.Second),
-				telemetry.WithLogger(log),
-				telemetry.WithDebug(rootFlags.TelemetryDebugMode),
-				telemetry.WithAuthToken(os.Getenv("WG_TELEMETRY_AUTH_TOKEN")),
-			)
-
-			// Send telemetry in a goroutine to not block the command
-			go func() {
-				err := TelemetryClient.Send(metrics)
-
-				// AddMetric the usage of the command immediately
-				if rootFlags.TelemetryDebugMode {
-					if err != nil {
-						log.Error("Could not send telemetry data", zap.Error(err))
-					} else {
-						log.Info("Telemetry data sent")
-					}
-				}
-			}()
-		}
+		}()
 
 		return nil
 	},
@@ -174,22 +210,28 @@ type BuildTimeConfig struct {
 // Execute adds all child commands to the root command and sets flags appropriately.
 // This is called by main.main(). It only needs to happen once to the rootCmd.
 func Execute(buildInfo node.BuildInfo, githubAuthDemo node.GitHubAuthDemo) {
+	// Instead of recover()'ing, use a boolean that we set to false
+	// at the of the function to keep track of whether we've panic()'ed
+	// or not. This allows us to set an appropriate exit code while
+	// showing the whole panic stack, which might be useful for fixing the
+	// problem.
+	didPanic := true
 	var err error
-
 	defer func() {
 		// In case of a panic or error we want to flush the telemetry data
-		if r := recover(); r != nil || err != nil {
-			FlushTelemetry()
+		FlushTelemetry()
+		if didPanic || err != nil {
+			if err != nil {
+				fmt.Fprintf(os.Stderr, "%s\n", err)
+			}
 			os.Exit(1)
-		} else {
-			FlushTelemetry()
-			os.Exit(0)
 		}
 	}()
 
 	BuildInfo = buildInfo
 	GitHubAuthDemo = githubAuthDemo
 	err = rootCmd.Execute()
+	didPanic = false
 }
 
 func FlushTelemetry() {