+++ packages/sdk/src/server/plugins/hooks.test.ts
@@ -51,6 +51,13 @@ export const getFastify = async (serverConfig: WunderGraphHooksAndServerConfig) 
 
 			configHash: '',
 			environmentIds: [],
+			enabledFeatures: {
+				apiCount: 0,
+				schemaExtension: false,
+				customJSONScalars: false,
+				customIntScalars: false,
+				customFloatScalars: false,
+			},
 		},
 		serverConfig,
 		gracefulShutdown: false,