+++ pkg/telemetry/data_sources.go
@@ -4,10 +4,6 @@ import (
 	"fmt"
 	"net"
 	"net/url"
-	"os"
-	"path/filepath"
-
-	"google.golang.org/protobuf/proto"
 
 	"github.com/wundergraph/wundergraph/pkg/loadvariable"
 	"github.com/wundergraph/wundergraph/pkg/wgpb"
@@ -90,13 +86,8 @@ func dataSourceMetric(dataSourceTag string, urlVariable *wgpb.ConfigurationVaria
 }
 
 func DataSourceMetrics(wunderGraphDir string) ([]*Metric, error) {
-	configFile := filepath.Join(wunderGraphDir, "generated", "wundergraph.wgconfig")
-	configData, err := os.ReadFile(configFile)
+	wgConfig, err := readWunderGraphConfig(wunderGraphDir)
 	if err != nil {
-		return nil, err
-	}
-	var wgConfig wgpb.WunderGraphConfiguration
-	if err := proto.Unmarshal(configData, &wgConfig); err != nil {
 		return nil, err
 	}
 	var metrics []*Metric