+++ packages/golang-client/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.6.14](https://github.com/wundergraph/wundergraph/compare/@wundergraph/golang-client@0.6.13...@wundergraph/golang-client@0.6.14) (2023-07-14)
+
+**Note:** Version bump only for package @wundergraph/golang-client
+
+## [0.6.13](https://github.com/wundergraph/wundergraph/compare/@wundergraph/golang-client@0.6.12...@wundergraph/golang-client@0.6.13) (2023-07-13)
+
+**Note:** Version bump only for package @wundergraph/golang-client
+
+## [0.6.12](https://github.com/wundergraph/wundergraph/compare/@wundergraph/golang-client@0.6.11...@wundergraph/golang-client@0.6.12) (2023-07-06)
+
+**Note:** Version bump only for package @wundergraph/golang-client
+
+## [0.6.11](https://github.com/wundergraph/wundergraph/compare/@wundergraph/golang-client@0.6.10...@wundergraph/golang-client@0.6.11) (2023-07-05)
+
+**Note:** Version bump only for package @wundergraph/golang-client
+
 ## [0.6.10](https://github.com/wundergraph/wundergraph/compare/@wundergraph/golang-client@0.6.9...@wundergraph/golang-client@0.6.10) (2023-06-22)
 
 **Note:** Version bump only for package @wundergraph/golang-client